Transformation of \(\Sigma^{(k)}\) matrix to \(\Gamma\) matrix

Transforms the \(\Sigma^{(k)}\) matrix from the definition of a Huesler--Reiss distribution to the corresponding \(\Gamma\) matrix.

Sigma2Gamma(S, k = 1, full = FALSE)

Arguments

S
Numeric \((d - 1) \times (d - 1)\) covariance matrix \(\Sigma^{(k)}\) from the definition of a Huesler--Reiss distribution. Numeric \(d \times d\) covariance matrix if full = TRUE, see full parameter.

k
Integer between 1 (the default value) and d. Indicates which matrix \(\Sigma^{(k)}\) is represented by S.

full
Logical. If true, then the kth row and column in \(\Sigma^{(k)}\) are included and the function returns a \(d \times d\) matrix. By default, full = FALSE.

Value

Numeric \(d \times d\) \(\Gamma\) matrix.

Details

For any k from 1 to d, the \(\Sigma^{(k)}\) matrix of size \((d - 1) \times (d - 1)\) in the definition of a Huesler--Reiss distribution can be transformed into a the corresponding \(d \times d\) \(\Gamma\) matrix. If full = TRUE, then \(\Sigma^{(k)}\) must be a \(d \times d\) matrix with kth row and column containing zeros. For details see Engelke and Hitz (2020) . This is the inverse of function of Gamma2Sigma.

References

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

Sigma1 <-  rbind(c(1.5, 0.5, 1),
                 c(0.5, 1.5, 1),
                 c(1, 1, 2))
Sigma2Gamma(Sigma1, k = 1, full = FALSE)

#>      [,1] [,2] [,3] [,4]
#> [1,]  0.0  1.5  1.5  2.0
#> [2,]  1.5  0.0  2.0  1.5
#> [3,]  1.5  2.0  0.0  1.5
#> [4,]  2.0  1.5  1.5  0.0