Completion of non-decomposable Gamma matrices

Given a graph and variogram matrix Gamma, returns the full Gamma matrix implied by the conditional independencies. This function uses a convergent iterative algorithm.

complete_Gamma_general(Gamma, graph, N = 1000, tol = 0, check_tol = 100)

Arguments

Gamma
A complete variogram matrix (without any graphical structure)

graph
An igraph::graph object

N
The maximal number of iterations of the algorithm

tol
The tolerance to use when checking for zero entries in Theta

check_tol
After how many iterations to check the tolerance in Theta

Value

A matrix that agrees with Gamma on the entries corresponding to edges in graph and the diagonals. The corresponding \(\Theta\) matrix produced by Gamma2Theta has values close to zero in the remaining entries (how close depends on the input and the number of iterations).

See also