DEMO-VERSION: Completion of non-decomposable Gamma matrices

Given a graph and variogram matrix Gamma, returns the full Gamma matrix implied by the conditional independencies. DEMO VERSION: Returns a lot of details and allows specifying the graph list that is used. Is way slower than other functions.

Usage

complete_Gamma_general_demo(Gamma, graph, N = 1000, tol = 0, gList = NULL)

Arguments

Gamma

A complete variogram matrix (without any graphical structure).

graph

An igraph::graph object.

N

The maximal number of iterations of the algorithm.

tol

The tolerance to use when checking for zero entries in Theta.

gList

A list of graphs to be used instead of the output from make_sep_list().

Value

A nested list, containing the following details. The "error term" is the maximal absolute value of Theta in a non-edge entry.

graph, N, tol

As in the input

gList

As in the input or computed by make_sep_list().

Gamma0, Theta0, err0

Initial Gamma, Theta, and error term.

iterations

A nested list, containing the following infos for each performed iteration:

n

Number of the iteration

t

Corresponding index in gList

g

The graph used

Gamma, Theta, err

The value of Gamma, Theta, and error term after the iteration

See also