Non-decomposable completion of variogram matrices

Given a non-decomposable graph, and (non-graphical) variogram matrix Gamma, modifies Gamma in non-edge entries, such that the resulting matrix is a variogram matrix with graphical structure described by graph. Does so by splitting graph at complete separators into smaller subgraphs, and calling complete_Gamma_general for each subgraph/submatrix, using multiple cores if available.

Usage

complete_Gamma_general_split(
  Gamma,
  graph,
  N = 10000,
  sub_tol = get_large_tol() * 0.001,
  check_tol = 100,
  mc_cores_overwrite = NULL,
  final_tol = get_large_tol()
)

Arguments

Gamma

Numeric \(d \times d\) variogram matrix.

graph

igraph::graph() object.

N

Maximum number of iterations.

sub_tol

Numeric scalar. Tolerance to be used when completing submatrices. Should be smaller than final_tol.

check_tol

Numeric/integer scalar. How often to check the tolerance when completing submatrices.

mc_cores_overwrite

NULL or numeric/integer scalar. Maximal number of cores to use.

final_tol

Numeric scalar. Check convergence of the final result with this tolerance. Skipped if this value is < 0.

Value

A completed \(d \times d\) variogram matrix.

See also