Upper Danube basin dataset

A dataset containing river discharge data for tributaries of the Danube.

Usage

danube

Format

A named list with four entries

data_clustered

A numeric matrix, containing pre-processed discharge data for each gauging station

data_raw

A numeric matrix, containing daily (raw) discharge data for each gauging station

info

A data frame, containing information about each gauging station

flow_edges

A two-column numeric matrix. Each row contains the indices (in info) of a pair of gauging stations that are directly connected by a river.

Source

Bavarian Environmental Agency https://www.gkd.bayern.de.

Details

To obtain the matrix data_clustered, the daily discharge data from the summer months of 1960 to 2010, given in data_raw, was declustered, yielding between seven and ten observations per year. Each row corresponds to one observation from this declustered time series, the non-unique rownames indicate which year an observation is from. Each column corresponds to one of the gauging stations, with column indices in data_raw/data_clustered corresponding to row indices in info. See asadi2015graphicalExtremes for details on the preprocessing and declustering.

info is a data frame containing the following information for each of the gauging stations or its corresponding catchment area:

RivNames

Name of the river at the gauging station

Lat, Long

Coordinates of the gauging station

Lat_Center, Long_Center

Coordinates of the center of the catchment corresponding to the gauging station

Alt

Mean altitude of the catchment

Area

Area of the catchment corresponding to the gauging station

Slope

Mean slope of the catchment

PlotCoordX, PlotCoordY

X-Y-coordinates which can be used to arrange the gauging stations when plotting a flow graph.

References

See also

Examples

dim(danube$data_clustered)
#> [1] 428  31
colnames(danube$info)
#>  [1] "RivNames"    "Lat"         "Long"        "AveVol"      "Lat_Center" 
#>  [6] "Long_Center" "Alt"         "Area"        "Chos"        "Density"    
#> [11] "Slope"       "PlotCoordX"  "PlotCoordY"