Learning extremal graph structure with latent variables

Following the methodology from engelkeTaeb2024;textualgraphicalExtremes, fits an extremal graph structure with latent variables.

Usage

eglatent(
  Gamma,
  lam1_list = c(0.1, 0.15, 0.19, 0.205),
  lam2_list = c(2),
  refit = TRUE,
  verbose = FALSE
)

Arguments

Gamma

conditionally negative semidefinite matrix. This will be typically the empirical variogram matrix.

lam1_list

Numeric vector of non-negative regularization parameters for eglatent. Default is lam1_list = c(0.1, 0.15, 0.19, 0.205).

lam2_list

Numeric vector of non-negative regularization parameters for eglatent. Default is lam2_list = c(2).

refit

Logical scalar, if TRUE then the model is refit on the estimated graph to obtain an estimate of the Gamma matrix on that graph. Default is refit = TRUE.

verbose

Logical scalar, indicating whether to print progress updates.

Value

The function fits one model for each combination of values in lam1_list and lam2_list. All returned objects have one entry per model. List consisting of: #'

graph

A list of igraph::graph objects representing the fitted graphs.

rk

Numeric vector containing the estimated ranks of the latent variables.

G_est

A list of numeric estimated variogram matrices corresponding to the fitted graphs.

G_refit

A list of numeric estimated variogram matrices refitted with fixed graphs corresponding to the fitted graphs.

lambdas

A list containing the values of lam1_list and lam2_list used for the model fit.

References

See also