Empirical estimation of extremal correlation

Estimates the d-dimensional extremal correlation coefficient empirically.

Usage

emp_chi_multdim(data, p = NULL)

Arguments

data

Numeric matrix, where n is the number of observations and d is the dimension.

p

Numeric scalar between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in data2mpareto() to standardize the data.

Value

Numeric scalar. The empirical d-dimensional extremal correlation coefficient for the data.

See also

Examples

n <- 100
d <- 2
p <- .8
G <- cbind(
  c(0, 1.5),
  c(1.5, 0)
)

set.seed(123)
my_data <- rmstable(n, "HR", d = d, par = G)
emp_chi_multdim(my_data, p)
#> [1] 0.55