Performs Gaussian likelihood optimization under Laplacian matrix constraints.

This function implements a block descent algorithm to find the maximum of the Gaussian log-likelihood under the constraint that the concentration matrix is a Laplacian matrix. See Röttger et al. (2021) for details.

Usage

emtp2(Gamma, tol = 1e-06, verbose = TRUE, initial_point = TRUE)

Arguments

Gamma

conditionally negative semidefinite matrix. This will be typically the empirical variogram matrix.

tol

The convergence tolerance. The algorithm terminates when the sum of absolute differences between two iterations is below tol.

verbose

if TRUE (default) the output will be printed.

initial_point

if TRUE (default), the algorithm will construct an initial point before the iteration steps.

Value

A list consisting of:

G_emtp2

The optimal value of the variogram matrix

it

The number of iterations

References

Röttger F, Engelke S, Zwiernik P (2021). “Total positivity in multivariate extremes.” doi:10.48550/ARXIV.2112.14727 , https://arxiv.org/abs/2112.14727.

See also