Experimental: Fit graph using empirical Theta matrix

Fits a graph to an empirical Gamma matrix by computing the corresponding Theta matrix using Gamma2Theta() and greedily chooses m edges that correspond to high absolute values in Theta.

Usage

fit_graph_to_Theta(data, m = NULL, Gamma_emp = NULL)

Arguments

data

The (standardized) data from which to compute Gamma

m

The number of edges to add, defaults to the number of edges in a tree

Gamma_emp

The empirical Gamma matrix (can be NULL if data is given)

Value

A list containing an [igraph::graph] object and a fitted Gamma matrix

See also