Parameter fitting for multivariate Huesler--Reiss Pareto distribution

Fits the parameters of a multivariate Huesler--Reiss Pareto distribution using (censored) likelihood estimation.

fmpareto_HR(
  data,
  p = NULL,
  cens = FALSE,
  init,
  maxit = 100,
  graph = NULL,
  method = "BFGS"
)

Arguments

data
Numeric matrix of size \(n\times d\), where \(n\) is the number of observations and \(d\) is the dimension.

p
Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto to standardize the data.

cens
Logical. If true, then censored likelihood contributions are used for components below the threshold. By default, cens = FALSE.

init
Numeric vector. Initial parameter values in the optimization. If graph is given, then the entries should correspond to the edges of the graph.

maxit
Positive integer. The maximum number of iterations in the optimization.

graph
Graph object from igraph package or NULL. If provided, the graph must be an undirected block graph, i.e., a decomposable, connected graph with singleton separator sets.

method
String. A valid optimization method used by the function optim. By default, method = "BFGS".

Value

List consisting of:

convergence: Logical. Indicates whether the optimization converged or not.

par: Numeric vector. Optimized parameters.

Gamma: Numeric matrix \(d \times d\). Fitted variogram matrix.

nllik: Numeric. Optimized value of the negative log-likelihood function.

hessian: Numeric matrix. Estimated Hessian matrix of the estimated parameters.

Details

If graph = NULL, then the parameters of a \(d \times d\) parameter matrix \(\Gamma\) of a Huesler--Reiss Pareto distribution are fitted. If graph is provided, then the conditional independence structure of this graph is assumed and the parameters on the edges are fitted. In both cases the full likelihood is used and therefore this function should only be used for small dimensions, say, \(d<5\). For models in higher dimensions fitting can be done separately on the cliques; see fmpareto_graph_HR.