Parameter fitting for Huesler-Reiss graphical models

Fits the parameter matrix (variogram) of a multivariate Huesler-Reiss Pareto distribution with a given graphical structure, using maximum-likelihood estimation or the empirical variogram.

Usage

fmpareto_graph_HR(
  data,
  graph,
  p = NULL,
  method = c("vario", "ML"),
  handleCliques = c("average", "full", "sequential"),
  ...
)

Arguments

data

Numeric matrix, where n is the number of observations and d is the number of dimensions.

graph

Undirected, connected [igraph::graph] object with d vertices, representing the graphical structure of the fitted Huesler-Reiss model.

p

Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data is already on a multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto() to standardize the data.

method

One of c('vario', 'ML'), with 'vario' as default, indicating the method to be used for parameter estimation. See details.

handleCliques

How to handle cliques and separators in the graph. See details.

...

Arguments passed to fmpareto_HR_MLE(). Currently cens, maxit, optMethod, and useTheta are supported.

Value

The estimated parameter matrix.

Details

If handleCliques='average', the marginal parameter matrix is estimated for each maximal clique of the graph and then combined into a partial parameter matrix by taking the average of entries from overlapping cliques. Lastly, the full parameter matrix is computed using complete_Gamma().

If handleCliques='full', first the full parameter matrix is estimated using the specified method and then the non-edge entries are adjusted such that the final parameter matrix has the graphical structure indicated by graph.

If handleCliques='sequential', graph must be decomposable, and method='ML' must be specified. The parameter matrix is first estimated on the (recursive) separators and then on the rest of the cliques, keeping previously estimated entries fixed.

If method='ML', the computational cost is mostly influenced by the total size of the graph (if handleCliques='full') or the size of the cliques, and can already take a significant amount of time for modest dimensions (e.g. d=3).

See also