Get Cliques and Separators of a graph

Finds all cliques, separators, and (recursively) separators of separators in a graph.

Usage

get_cliques_and_separators(
  graph,
  sortIntoLayers = FALSE,
  includeSingletons = FALSE
)

Arguments

graph

An [igraph::graph] object

Value

A list of vertex sets that represent the cliques and (recursive) separators of graph, ordered such that separators come before cliques they separate.