Number of cores to be used in parallel computations

Helper function that returns the number of cores to be used in parallel computations. Will always be 1 on Windows. On other systems, this value can be set using setOption('graphicalExtremes.mc.cores', ...).

Usage

get_mc_cores(overwrite = NULL)

Arguments

overwrite

Use this value (if it is valid and not on Windows)

Value

An integer to be used as number of cores

See also