Compute the exponent measure density of HR distribution

Computes the exponent measure density of HR distribution.

Usage

logdV_HR(x, Gamma = NULL, Theta = NULL)

Arguments

x

Numeric matrix or vector with d elements.

Gamma

d x d variogram matrix or numeric vector with d(d-1)/2 elements, containing the upper triangular part of a variogram matrix.

Theta

d x d precision matrix or numeric vector with d(d-1)/2 elements, containing the upper triangular part of a precision matrix.

Value

Numeric. The censored exponent measure of the HR distribution.

Details

Both Gamma and Theta are needed internally, but if one is missing it is computed from the other one.