Compute Huesler--Reiss log-likelihood, AIC, and BIC

Computes (censored) Huesler--Reiss log-likelihood, AIC, and BIC values.

Usage

loglik_HR(data, p = NULL, graph, Gamma, cens = FALSE)

Arguments

data

Numeric matrix \(n\times d\). It contains observations following a multivariate HR Pareto distribution.

p

Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto to standardize the data.

graph

An igraph::graph object. The graph must be undirected and connected

Gamma

Numeric matrix \(n\times d\). It represents a variogram matrix \(\Gamma\).

cens

Boolean. If true, then censored log-likelihood is computed. By default, cens = FALSE.

Value

Numeric vector c("loglik"=..., "aic"=..., "bic"=...) with the evaluated log-likelihood, AIC, and BIC values.