Simulate Dirichlet extremal functions on a tree

Simulates Dirichlet extremal functions on a tree

simu_px_tree_dirichlet(n, alpha.start, alpha.end, A)

Arguments

n
Number of simulations.

alpha.start
Numeric vector with \(d - 1\) elements, where \(d\) is the number of nodes in the tree (and \(d - 1\) is the number of edges).

alpha.end
Numeric vector with \(d - 1\) elements, where \(d\) is the number of nodes in the tree (and \(d - 1\) is the number of edges).

A
Numeric matrix \(d \times (d - 1)\); the rows represent the nodes in the tree, the columns represent the edges. For a fixed node \(k = 1, \dots, d\), each entry \((i, j)\) is equal to 1 if the edge in position \(j\) is on the directed path from node \(k\) to node \(i\) in the polytree rooted at node \(k\).

Value

Numeric matrix \(n\times d\). Simulated data.