Data standardization to multivariate Pareto scale

Transforms the data matrix empirically to the multivariate Pareto scale.

Usage

data2mpareto(data, p, na.rm = FALSE)

Arguments

data

Numeric \(n \times d\) matrix, where n is the number of observations and d is the dimension.

p

Numeric between 0 and 1. Probability used for the quantile to threshold the data.

na.rm

Logical. If rows containing NAs should be removed.

Value

Numeric \(m \times d\) matrix, where m is the number of rows in the original data matrix that are above the threshold.

Details

The columns of the data matrix are first transformed empirically to standard Pareto distributions. Then, only the observations where at least one component exceeds the p-quantile of the standard Pareto distribution are kept. Those observations are finally divided by the p-quantile of the standard Pareto distribution to standardize them to the multivariate Pareto scale.

If na.rm is FALSE, missing entries are left as such during the transformation of univariate marginals. In the thresholding step, missing values are considered as -Inf.

See also

Examples

n <- 20
d <- 4
p <- .8
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

set.seed(123)
my_data <- rmstable(n, "HR", d = d, par = G)
data2mpareto(my_data, p)
#>            [,1]      [,2]      [,3]      [,4]
#>  [1,] 0.2470588 0.4200000 1.4000000 0.7000000
#>  [2,] 1.0500000 0.6000000 0.3500000 0.2333333
#>  [3,] 0.6000000 0.3230769 1.0500000 0.3000000
#>  [4,] 0.7000000 0.5250000 0.8400000 1.4000000
#>  [5,] 2.1000000 1.4000000 0.6000000 0.6000000
#>  [6,] 0.4666667 0.3000000 0.2210526 2.1000000
#>  [7,] 0.3818182 2.1000000 0.3818182 1.0500000
#>  [8,] 1.4000000 0.8400000 4.2000000 0.4666667
#>  [9,] 4.2000000 1.0500000 2.1000000 4.2000000
#> [10,] 0.3230769 4.2000000 0.2100000 0.2470588