Learning extremal graph structure

Following the methodology from Engelke et al. (2022) , fits an extremal graph structure using the neighborhood selection approach (see Meinshausen and Bühlmann (2006) ) or graphical lasso (see Friedman et al. (2008) ).

Usage

eglearn(
  data,
  p = NULL,
  rholist = c(0.1, 0.15, 0.19, 0.205),
  reg_method = c("ns", "glasso"),
  complete_Gamma = FALSE
)

Arguments

data

Numeric \(n \times d\) matrix, where n is the number of observations and d is the dimension.

p

Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto() to standardize the data.

rholist

Numeric vector of non-negative regularization parameters for the lasso. Default is rholist = c(0.1, 0.15, 0.19, 0.205). For details see glasso::glassopath().

reg_method

One of "ns", "glasso", for neighborhood selection and graphical lasso, respectively. Default is reg_method = "ns". For details see Meinshausen and Bühlmann (2006) , Friedman et al. (2008) .

complete_Gamma

Whether you want to try fto complete Gamma matrix. Default is complete_Gamma = FALSE.

Value

List made of:

graph

A list of [igraph::graph] objects representing the fitted graphs for each rho in rholist.

Gamma

A list of numeric estimated \(d \times d\) variogram matrices \(\Gamma\) corresponding to the fitted graphs, for each rho in rholist. If complete_Gamma = FALSE or the underlying graph is not connected, it returns NULL.

rholist

The list of penalty coefficients.

graph_ic

A list of [igraph::graph] objects representing the optimal graph according to the aic, bic, and mbic information criteria. If reg_method = "glasso", it returns a list of NULL.

Gamma_ic

A list of numeric \(d \times d\) estimated variogram matrices \(\Gamma\) corresponding to the aic, bic, and mbic information criteria. If reg_method = "glasso", complete_Gamma = FALSE, or the underlying graph is not connected, it returns a list of NULL.

References

Engelke S, Lalancette M, Volgushev S (2022). “Learning extremal graphical structures in high dimensions.” doi:10.48550/ARXIV.2111.00840 , Available from https://arxiv.org/abs/2111.00840., 2111.00840, https://arxiv.org/abs/2111.00840.

Friedman J, Hastie T, Tibshirani R (2008). “Sparse inverse covariance estimation with the graphical lasso.” Biostatistics, 9(3), 432--441.

Meinshausen N, Bühlmann P (2006). “High-dimensional graphs and variable selection with the Lasso.” The Annals of Statistics, 34(3), 1436 -- 1462. doi:10.1214/009053606000000281 .

See also