Estimation of the variogram matrix \(\Gamma\) of a Huesler-Reiss distribution

Estimates the variogram of the Huesler-Reiss distribution empirically.

Usage

emp_vario(data, k = NULL, p = NULL)

emp_vario_pairwise(data, k = NULL, p = NULL, verbose = FALSE)

Arguments

data

Numeric \(n \times d\) matrix, where n is the number of observations and d is the dimension.

k

Integer between 1 and d. Component of the multivariate observations that is conditioned to be larger than the threshold p. If NULL (default), then an average over all k is returned.

p

Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto() to standardize the data.

verbose

Print verbose progress information

Value

Numeric \(d \times d\) matrix. The estimated variogram of the Huesler-Reiss distribution.

Details

emp_vario_pairwise calls emp_vario for each pair of observations. This is more robust if the data contains many NAs, but can take rather long.

See also