Parameter fitting for multivariate Huesler--Reiss Pareto distributions on block graphs

Fits the parameters of a multivariate Huesler--Reiss Pareto distribution using (censored) likelihood estimation. See Engelke and Hitz (2020) for details.

Usage

fmpareto_graph_HR(
  data,
  graph,
  p = NULL,
  method = c("ML", "vario"),
  cens = FALSE
)

Arguments

data

Numeric matrix of size \(n\times d\), where \(n\) is the number of observations and \(d\) is the dimension.

graph

Undirected graph object from igraph package.

p

Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto to standardize the data.

method

One of "ML", "vario". Default is method = "ML".

cens

Logical. If true, then censored likelihood contributions are used for components below the threshold. By default, cens = FALSE.

Value

List consisting of:

graph: An igraph::graph() object.

Gamma: Numeric \(d\times d\) estimated variogram matrix \(\Gamma\).

References

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

## Fitting a 4-dimensional HR distribution

my_graph <- igraph::graph_from_adjacency_matrix(
  rbind(
    c(0, 1, 0, 0),
    c(1, 0, 1, 1),
    c(0, 1, 0, 0),
    c(0, 1, 0, 0)
  ),
  mode = "undirected"
)
n <- 100
Gamma_vec <- c(.5, 1.4, .8)
complete_Gamma(Gamma = Gamma_vec, graph = my_graph) ## full Gamma matrix
#>      [,1] [,2] [,3] [,4]
#> [1,]  0.0  0.5  1.9  1.3
#> [2,]  0.5  0.0  1.4  0.8
#> [3,]  1.9  1.4  0.0  2.2
#> [4,]  1.3  0.8  2.2  0.0
edges_to_add <- rbind(c(1, 3), c(1, 4), c(3, 4))

set.seed(123)
my_data <- rmpareto_tree(n, "HR", tree = my_graph, par = Gamma_vec)
my_fit <- fmpareto_graph_HR(my_data,
  graph = my_graph, method = "ML",
  p = NULL, cens = FALSE
)