Parameter fitting for multivariate Huesler--Reiss Pareto distributions on block graphs

Fits the parameters of a multivariate Huesler--Reiss Pareto distribution using (censored) likelihood estimation. Fitting is done separately on the cliques of the block graph. If edges_to_add are provided, then these edges are added in a greedy search to the original graph, such that in each step the likelihood is improved maximally and the new graph stays in the class of block graphs. See Engelke and Hitz (2020) for details.

fmpareto_graph_HR(data, graph, p = NULL, cens = FALSE, edges_to_add = NULL)

Arguments

data
Numeric matrix of size \(n\times d\), where \(n\) is the number of observations and \(d\) is the dimension.

graph
Graph object from igraph package. The graph must be an undirected block graph, i.e., a decomposable, connected graph with singleton separator sets.

p
Numeric between 0 and 1 or NULL. If NULL (default), it is assumed that the data are already on multivariate Pareto scale. Else, p is used as the probability in the function data2mpareto to standardize the data.

cens
Logical. If true, then censored likelihood contributions are used for components below the threshold. By default, cens = FALSE.

edges_to_add
Numeric matrix \(m\times 2\), where \(m\) is the number of edges that are tried to be added in the greedy search. By default, edges_to_add = NULL.

Value

List consisting of:

graph: Graph object from igraph package. If edges_to_add are provided, then this is a list of the resulting graphs in each step of the greedy search.

Gamma: Numeric \(d\times d\) estimated variogram matrix \(\Gamma\). If edges_to_add are provided, then this is a list of the estimated variogram matrices in each step of the greedy search.

AIC: (only if edges_to_add are provided) List of AIC values of the fitted models in each step of the greedy search.

edges_added: (only if edges_to_add are provided) Numeric matrix \(m'\times 2\), where the \(m'\leq m\) rows contain the edges that were added in the greedy search.

References

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

## Fitting a 4-dimensional HR distribution

my_graph <- igraph::graph_from_adjacency_matrix(
  rbind(
    c(0, 1, 0, 0),
    c(1, 0, 1, 1),
    c(0, 1, 0, 0),
    c(0, 1, 0, 0)
  ),
  mode = "undirected"
)
n <- 100
Gamma_vec <- c(.5, 1.4, .8)
complete_Gamma(Gamma = Gamma_vec, graph = my_graph) ## full Gamma matrix

#>      [,1] [,2] [,3] [,4]
#> [1,]  0.0  0.5  1.9  1.3
#> [2,]  0.5  0.0  1.4  0.8
#> [3,]  1.9  1.4  0.0  2.2
#> [4,]  1.3  0.8  2.2  0.0

edges_to_add <- rbind(c(1, 3), c(1, 4), c(3, 4))

set.seed(123)
my_data <- rmpareto_tree(n, "HR", tree = my_graph, par = Gamma_vec)
my_fit <- fmpareto_graph_HR(my_data,
  graph = my_graph,
  p = NULL, cens = FALSE, edges_to_add = edges_to_add
)

#> Error in stats::optim(init_opt, nllik, hessian = TRUE, control = list(maxit = maxit),     method = method): non-finite value supplied by optim