Generate random Hüsler--Reiss Models

Generates a random connected graph and Gamma matrix with conditional independence structure corresponding to that graph.

generate_random_model(d, graph_type = "tree", ...)

Arguments

d
Number of vertices in the graph

graph_type
"tree", "block", "decomposable", or "general"

...
Further arguments passed to functions generating the graph and Gamma matrix

See also

Examples

set.seed(1)
d <- 12

generate_random_model(d, 'tree')

#> $graph
#> IGRAPH f0b0edf U--- 12 11 -- 
#> + edges from f0b0edf:
#>  [1] 1-- 7 1--12 2-- 3 3-- 8 3-- 9 4-- 5 4-- 6 6--10 7--10 7--11 9--10
#> 
#> $Gamma
#>            [,1]      [,2]      [,3]       [,4]       [,5]      [,6]      [,7]
#>  [1,] 0.0000000 2.6720780 2.0022634 1.75490900 1.84728711 1.5147334 0.4817166
#>  [2,] 2.6720780 0.0000000 0.6698146 1.80053975 1.89291786 1.5603642 2.1903614
#>  [3,] 2.0022634 0.6698146 0.0000000 1.13072516 1.22310327 0.8905496 1.5205468
#>  [4,] 1.7549090 1.8005397 1.1307252 0.00000000 0.09237811 0.2401756 1.2731924
#>  [5,] 1.8472871 1.8929179 1.2231033 0.09237811 0.00000000 0.3325537 1.3655705
#>  [6,] 1.5147334 1.5603642 0.8905496 0.24017559 0.33255370 0.0000000 1.0330168
#>  [7,] 0.4817166 2.1903614 1.5205468 1.27319238 1.36557049 1.0330168 0.0000000
#>  [8,] 2.3459810 1.0135322 0.3437176 1.47444277 1.56682088 1.2342672 1.8642644
#>  [9,] 1.8552114 0.8168666 0.1470520 0.98367317 1.07605128 0.7434976 1.3734948
#> [10,] 1.3132236 1.3588544 0.6890398 0.44168538 0.53406349 0.2015098 0.8315070
#> [11,] 1.2638652 2.9725099 2.3026953 2.05534092 2.14771903 1.8151653 0.7821485
#> [12,] 1.0796174 3.7516954 3.0818808 2.83452640 2.92690451 2.5943508 1.5613340
#>            [,8]      [,9]     [,10]     [,11]    [,12]
#>  [1,] 2.3459810 1.8552114 1.3132236 1.2638652 1.079617
#>  [2,] 1.0135322 0.8168666 1.3588544 2.9725099 3.751695
#>  [3,] 0.3437176 0.1470520 0.6890398 2.3026953 3.081881
#>  [4,] 1.4744428 0.9836732 0.4416854 2.0553409 2.834526
#>  [5,] 1.5668209 1.0760513 0.5340635 2.1477190 2.926905
#>  [6,] 1.2342672 0.7434976 0.2015098 1.8151653 2.594351
#>  [7,] 1.8642644 1.3734948 0.8315070 0.7821485 1.561334
#>  [8,] 0.0000000 0.4907696 1.0327574 2.6464129 3.425598
#>  [9,] 0.4907696 0.0000000 0.5419878 2.1556433 2.934829
#> [10,] 1.0327574 0.5419878 0.0000000 1.6136555 2.392841
#> [11,] 2.6464129 2.1556433 1.6136555 0.0000000 2.343483
#> [12,] 3.4255984 2.9348288 2.3928410 2.3434826 0.000000
#> 

generate_random_model(d, 'block')

#> $graph
#> IGRAPH c316b21 U--- 12 21 -- 
#> + edges from c316b21:
#>  [1]  1-- 2  2-- 3  2-- 4  2-- 5  2-- 6  2-- 7  2-- 8  2-- 9  2--10  4-- 5
#> [11]  4-- 6  5-- 6  7-- 8  7-- 9  7--10  7--11  7--12  8-- 9  8--10  9--10
#> [21] 11--12
#> 
#> $Gamma
#>            [,1]       [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.38002970 0.5128718 3.7521179 2.4857676 0.9578492 0.5603988
#>  [2,] 0.3800297 0.00000000 0.1328421 3.3720882 2.1057379 0.5778195 0.1803691
#>  [3,] 0.5128718 0.13284212 0.0000000 3.5049303 2.2385800 0.7106617 0.3132112
#>  [4,] 3.7521179 3.37208816 3.5049303 0.0000000 0.4559447 1.8011275 3.5524572
#>  [5,] 2.4857676 2.10573786 2.2385800 0.4559447 0.0000000 1.2910493 2.2861069
#>  [6,] 0.9578492 0.57781954 0.7106617 1.8011275 1.2910493 0.0000000 0.7581886
#>  [7,] 0.5603988 0.18036908 0.3132112 3.5524572 2.2861069 0.7581886 0.0000000
#>  [8,] 0.5067318 0.12670214 0.2595443 3.4987903 2.2324400 0.7045217 0.1053878
#>  [9,] 0.4699951 0.08996541 0.2228075 3.4620536 2.1957033 0.6677850 0.1736513
#> [10,] 0.5444061 0.16437641 0.2972185 3.5364646 2.2701143 0.7421960 0.2060129
#> [11,] 0.8742990 0.49426927 0.6271114 3.8663574 2.6000071 1.0720888 0.3139002
#> [12,] 1.3181809 0.93815122 1.0709933 4.3102394 3.0438891 1.5159708 0.7577821
#>             [,8]       [,9]     [,10]     [,11]     [,12]
#>  [1,] 0.50673183 0.46999511 0.5444061 0.8742990 1.3181809
#>  [2,] 0.12670214 0.08996541 0.1643764 0.4942693 0.9381512
#>  [3,] 0.25954425 0.22280753 0.2972185 0.6271114 1.0709933
#>  [4,] 3.49879029 3.46205357 3.5364646 3.8663574 4.3102394
#>  [5,] 2.23243999 2.19570327 2.2701143 2.6000071 3.0438891
#>  [6,] 0.70452168 0.66778496 0.7421960 1.0720888 1.5159708
#>  [7,] 0.10538784 0.17365132 0.2060129 0.3139002 0.7577821
#>  [8,] 0.00000000 0.07218944 0.1525241 0.4192880 0.8631700
#>  [9,] 0.07218944 0.00000000 0.1396240 0.4875515 0.9314335
#> [10,] 0.15252407 0.13962401 0.0000000 0.5199131 0.9637950
#> [11,] 0.41928804 0.48755151 0.5199131 0.0000000 0.2993314
#> [12,] 0.86316999 0.93143346 0.9637950 0.2993314 0.0000000
#> 

generate_random_model(d, 'decomposable')

#> $graph
#> IGRAPH c830878 U--- 12 25 -- 
#> + edges from c830878:
#>  [1]  1-- 2  1-- 3  1-- 7  1-- 8  2-- 3  2-- 4  2-- 5  2-- 6  2-- 7  2-- 8
#> [11]  2-- 9  2--10  2--11  4-- 5  4-- 6  4-- 9  4--10  4--11  4--12  5-- 6
#> [21]  5--12  7-- 8  9--10  9--11 10--11
#> 
#> $Gamma
#>            [,1]      [,2]      [,3]       [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.2072024 0.1386426 0.45955830 0.4605872 0.2945035 0.4398021
#>  [2,] 0.2072024 0.0000000 0.2646398 0.25235592 0.2533848 0.0873011 0.1755466
#>  [3,] 0.1386426 0.2646398 0.0000000 0.51699569 0.5180246 0.3519409 0.4860596
#>  [4,] 0.4595583 0.2523559 0.5169957 0.00000000 0.6699109 0.1489949 0.4279025
#>  [5,] 0.4605872 0.2533848 0.5180246 0.66991088 0.0000000 0.3154622 0.4289314
#>  [6,] 0.2945035 0.0873011 0.3519409 0.14899488 0.3154622 0.0000000 0.2628477
#>  [7,] 0.4398021 0.1755466 0.4860596 0.42790253 0.4289314 0.2628477 0.0000000
#>  [8,] 0.3328008 0.6838590 0.4996328 0.93621489 0.9372438 0.7711601 0.7628150
#>  [9,] 1.2633699 1.0561675 1.3208073 0.38677108 1.6093753 0.7952634 1.2317141
#> [10,] 0.8844230 0.6772206 0.9418604 0.25448673 1.1501952 0.5094968 0.8527672
#> [11,] 0.4985067 0.2913044 0.5559441 0.09057542 0.6920664 0.2074462 0.4668510
#> [12,] 1.1002993 0.8930969 1.1577367 0.38261960 1.5695674 0.7258932 1.0686435
#>            [,8]      [,9]     [,10]      [,11]     [,12]
#>  [1,] 0.3328008 1.2633699 0.8844230 0.49850674 1.1002993
#>  [2,] 0.6838590 1.0561675 0.6772206 0.29130436 0.8930969
#>  [3,] 0.4996328 1.3208073 0.9418604 0.55594414 1.1577367
#>  [4,] 0.9362149 0.3867711 0.2544867 0.09057542 0.3826196
#>  [5,] 0.9372438 1.6093753 1.1501952 0.69206636 1.5695674
#>  [6,] 0.7711601 0.7952634 0.5094968 0.20744615 0.7258932
#>  [7,] 0.7628150 1.2317141 0.8527672 0.46685097 1.0686435
#>  [8,] 0.0000000 1.7400265 1.3610796 0.97516334 1.5769559
#>  [9,] 1.7400265 0.0000000 0.2040745 0.44858546 0.5561065
#> [10,] 1.3610796 0.2040745 0.0000000 0.38005204 0.5499711
#> [11,] 0.9751633 0.4485855 0.3800520 0.00000000 0.4995983
#> [12,] 1.5769559 0.5561065 0.5499711 0.49959825 0.0000000
#> 

generate_random_model(d, 'general')

#> $graph
#> IGRAPH 495620b U--- 12 24 -- Erdos renyi (gnm) graph
#> + attr: name (g/c), type (g/c), loops (g/l), m (g/n)
#> + edges from 495620b:
#>  [1]  1-- 4  1-- 5  3-- 5  4-- 5  2-- 6  1-- 7  2-- 8  4-- 8  5-- 8  1-- 9
#> [11]  6-- 9  8-- 9  1--10  3--10  4--10  5--10  7--10  1--11  4--11  8--11
#> [21] 10--11  1--12  7--12  9--12
#> 
#> $Gamma
#>             [,1]       [,2]      [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.00000000 0.24140512 0.3653153 0.08010552 0.06002560 0.47971068
#>  [2,] 0.24140512 0.00000000 0.4259775 0.16403568 0.25658949 0.31463187
#>  [3,] 0.36531528 0.42597747 0.0000000 0.26328189 0.34279470 0.68751983
#>  [4,] 0.08010552 0.16403568 0.2632819 0.00000000 0.08380705 0.42266129
#>  [5,] 0.06002560 0.25658949 0.3427947 0.08380705 0.00000000 0.50067464
#>  [6,] 0.47971068 0.31463187 0.6875198 0.42266129 0.50067464 0.00000000
#>  [7,] 0.15412273 0.33612113 0.3724520 0.17495155 0.19121187 0.58178213
#>  [8,] 0.18658659 0.06234987 0.3674563 0.10597928 0.20085002 0.34792026
#>  [9,] 0.48613569 0.32448581 0.6949887 0.42999917 0.50735930 0.01804742
#> [10,] 0.27700054 0.31866623 0.1467762 0.15738060 0.25972245 0.58264878
#> [11,] 0.12293919 0.15365178 0.3074049 0.04876885 0.13164636 0.41916637
#> [12,] 0.11012073 0.34347821 0.4692875 0.18654935 0.16895139 0.57304614
#>            [,7]       [,8]       [,9]     [,10]      [,11]     [,12]
#>  [1,] 0.1541227 0.18658659 0.48613569 0.2770005 0.12293919 0.1101207
#>  [2,] 0.3361211 0.06234987 0.32448581 0.3186662 0.15365178 0.3434782
#>  [3,] 0.3724520 0.36745630 0.69498874 0.1467762 0.30740491 0.4692875
#>  [4,] 0.1749516 0.10597928 0.42999917 0.1573806 0.04876885 0.1865493
#>  [5,] 0.1912119 0.20085002 0.50735930 0.2597224 0.13164636 0.1689514
#>  [6,] 0.5817821 0.34792026 0.01804742 0.5826488 0.41916637 0.5730461
#>  [7,] 0.0000000 0.28013056 0.58853758 0.2667975 0.21799176 0.2558738
#>  [8,] 0.2801306 0.00000000 0.35646863 0.2597562 0.09449767 0.2900520
#>  [9,] 0.5885376 0.35646863 0.00000000 0.5902273 0.42681373 0.5790786
#> [10,] 0.2667975 0.25975624 0.59022732 0.0000000 0.20098416 0.3802803
#> [11,] 0.2179918 0.09449767 0.42681373 0.2009842 0.00000000 0.2286348
#> [12,] 0.2558738 0.29005195 0.57907861 0.3802803 0.22863476 0.0000000
#>