Generate random Hüsler--Reiss Models

Generates a random connected graph and Gamma matrix with conditional independence structure corresponding to that graph.

Usage

generate_random_model(d, graph_type = "tree", ...)

Arguments

d

Number of vertices in the graph

graph_type

"tree", "block", "decomposable", or "general"

...

Further arguments passed to functions generating the graph and Gamma matrix

See also

Examples

set.seed(1)
d <- 12

generate_random_model(d, 'tree')
#> $graph
#> IGRAPH 5a90bbd U--- 12 11 -- 
#> + edges from 5a90bbd:
#>  [1] 1-- 7 1--12 2-- 3 3-- 8 3-- 9 4-- 5 4-- 6 6--10 7--10 7--11 9--10
#> 
#> $Gamma
#>            [,1]      [,2]      [,3]       [,4]       [,5]      [,6]      [,7]
#>  [1,] 0.0000000 2.6720780 2.0022634 1.75490900 1.84728711 1.5147334 0.4817166
#>  [2,] 2.6720780 0.0000000 0.6698146 1.80053975 1.89291786 1.5603642 2.1903614
#>  [3,] 2.0022634 0.6698146 0.0000000 1.13072516 1.22310327 0.8905496 1.5205468
#>  [4,] 1.7549090 1.8005397 1.1307252 0.00000000 0.09237811 0.2401756 1.2731924
#>  [5,] 1.8472871 1.8929179 1.2231033 0.09237811 0.00000000 0.3325537 1.3655705
#>  [6,] 1.5147334 1.5603642 0.8905496 0.24017559 0.33255370 0.0000000 1.0330168
#>  [7,] 0.4817166 2.1903614 1.5205468 1.27319238 1.36557049 1.0330168 0.0000000
#>  [8,] 2.3459810 1.0135322 0.3437176 1.47444277 1.56682088 1.2342672 1.8642644
#>  [9,] 1.8552114 0.8168666 0.1470520 0.98367317 1.07605128 0.7434976 1.3734948
#> [10,] 1.3132236 1.3588544 0.6890398 0.44168538 0.53406349 0.2015098 0.8315070
#> [11,] 1.2638652 2.9725099 2.3026953 2.05534092 2.14771903 1.8151653 0.7821485
#> [12,] 1.0796174 3.7516954 3.0818808 2.83452640 2.92690451 2.5943508 1.5613340
#>            [,8]      [,9]     [,10]     [,11]    [,12]
#>  [1,] 2.3459810 1.8552114 1.3132236 1.2638652 1.079617
#>  [2,] 1.0135322 0.8168666 1.3588544 2.9725099 3.751695
#>  [3,] 0.3437176 0.1470520 0.6890398 2.3026953 3.081881
#>  [4,] 1.4744428 0.9836732 0.4416854 2.0553409 2.834526
#>  [5,] 1.5668209 1.0760513 0.5340635 2.1477190 2.926905
#>  [6,] 1.2342672 0.7434976 0.2015098 1.8151653 2.594351
#>  [7,] 1.8642644 1.3734948 0.8315070 0.7821485 1.561334
#>  [8,] 0.0000000 0.4907696 1.0327574 2.6464129 3.425598
#>  [9,] 0.4907696 0.0000000 0.5419878 2.1556433 2.934829
#> [10,] 1.0327574 0.5419878 0.0000000 1.6136555 2.392841
#> [11,] 2.6464129 2.1556433 1.6136555 0.0000000 2.343483
#> [12,] 3.4255984 2.9348288 2.3928410 2.3434826 0.000000
#> 
generate_random_model(d, 'block')
#> $graph
#> IGRAPH 0d27518 U--- 12 21 -- 
#> + edges from 0d27518:
#>  [1]  1-- 2  2-- 3  2-- 4  2-- 5  2-- 6  2-- 7  2-- 8  2-- 9  2--10  4-- 5
#> [11]  4-- 6  5-- 6  7-- 8  7-- 9  7--10  7--11  7--12  8-- 9  8--10  9--10
#> [21] 11--12
#> 
#> $Gamma
#>            [,1]       [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.38002970 0.5128718 3.7521179 2.4857676 0.9578492 0.5603988
#>  [2,] 0.3800297 0.00000000 0.1328421 3.3720882 2.1057379 0.5778195 0.1803691
#>  [3,] 0.5128718 0.13284212 0.0000000 3.5049303 2.2385800 0.7106617 0.3132112
#>  [4,] 3.7521179 3.37208816 3.5049303 0.0000000 0.4559447 1.8011275 3.5524572
#>  [5,] 2.4857676 2.10573786 2.2385800 0.4559447 0.0000000 1.2910493 2.2861069
#>  [6,] 0.9578492 0.57781954 0.7106617 1.8011275 1.2910493 0.0000000 0.7581886
#>  [7,] 0.5603988 0.18036908 0.3132112 3.5524572 2.2861069 0.7581886 0.0000000
#>  [8,] 0.5067318 0.12670214 0.2595443 3.4987903 2.2324400 0.7045217 0.1053878
#>  [9,] 0.4699951 0.08996541 0.2228075 3.4620536 2.1957033 0.6677850 0.1736513
#> [10,] 0.5444061 0.16437641 0.2972185 3.5364646 2.2701143 0.7421960 0.2060129
#> [11,] 0.8742990 0.49426927 0.6271114 3.8663574 2.6000071 1.0720888 0.3139002
#> [12,] 1.3181809 0.93815122 1.0709933 4.3102394 3.0438891 1.5159708 0.7577821
#>             [,8]       [,9]     [,10]     [,11]     [,12]
#>  [1,] 0.50673183 0.46999511 0.5444061 0.8742990 1.3181809
#>  [2,] 0.12670214 0.08996541 0.1643764 0.4942693 0.9381512
#>  [3,] 0.25954425 0.22280753 0.2972185 0.6271114 1.0709933
#>  [4,] 3.49879029 3.46205357 3.5364646 3.8663574 4.3102394
#>  [5,] 2.23243999 2.19570327 2.2701143 2.6000071 3.0438891
#>  [6,] 0.70452168 0.66778496 0.7421960 1.0720888 1.5159708
#>  [7,] 0.10538784 0.17365132 0.2060129 0.3139002 0.7577821
#>  [8,] 0.00000000 0.07218944 0.1525241 0.4192880 0.8631700
#>  [9,] 0.07218944 0.00000000 0.1396240 0.4875515 0.9314335
#> [10,] 0.15252407 0.13962401 0.0000000 0.5199131 0.9637950
#> [11,] 0.41928804 0.48755151 0.5199131 0.0000000 0.2993314
#> [12,] 0.86316999 0.93143346 0.9637950 0.2993314 0.0000000
#> 
generate_random_model(d, 'decomposable')
#> $graph
#> IGRAPH 1fafbf5 U--- 12 25 -- 
#> + edges from 1fafbf5:
#>  [1]  1-- 2  1-- 3  1-- 7  1-- 8  2-- 3  2-- 4  2-- 5  2-- 6  2-- 7  2-- 8
#> [11]  2-- 9  2--10  2--11  4-- 5  4-- 6  4-- 9  4--10  4--11  4--12  5-- 6
#> [21]  5--12  7-- 8  9--10  9--11 10--11
#> 
#> $Gamma
#>            [,1]      [,2]      [,3]       [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.2072024 0.1386426 0.45955830 0.4605872 0.2945035 0.4398021
#>  [2,] 0.2072024 0.0000000 0.2646398 0.25235592 0.2533848 0.0873011 0.1755466
#>  [3,] 0.1386426 0.2646398 0.0000000 0.51699569 0.5180246 0.3519409 0.4860596
#>  [4,] 0.4595583 0.2523559 0.5169957 0.00000000 0.6699109 0.1489949 0.4279025
#>  [5,] 0.4605872 0.2533848 0.5180246 0.66991088 0.0000000 0.3154622 0.4289314
#>  [6,] 0.2945035 0.0873011 0.3519409 0.14899488 0.3154622 0.0000000 0.2628477
#>  [7,] 0.4398021 0.1755466 0.4860596 0.42790253 0.4289314 0.2628477 0.0000000
#>  [8,] 0.3328008 0.6838590 0.4996328 0.93621489 0.9372438 0.7711601 0.7628150
#>  [9,] 1.2633699 1.0561675 1.3208073 0.38677108 1.6093753 0.7952634 1.2317141
#> [10,] 0.8844230 0.6772206 0.9418604 0.25448673 1.1501952 0.5094968 0.8527672
#> [11,] 0.4985067 0.2913044 0.5559441 0.09057542 0.6920664 0.2074462 0.4668510
#> [12,] 1.1002993 0.8930969 1.1577367 0.38261960 1.5695674 0.7258932 1.0686435
#>            [,8]      [,9]     [,10]      [,11]     [,12]
#>  [1,] 0.3328008 1.2633699 0.8844230 0.49850674 1.1002993
#>  [2,] 0.6838590 1.0561675 0.6772206 0.29130436 0.8930969
#>  [3,] 0.4996328 1.3208073 0.9418604 0.55594414 1.1577367
#>  [4,] 0.9362149 0.3867711 0.2544867 0.09057542 0.3826196
#>  [5,] 0.9372438 1.6093753 1.1501952 0.69206636 1.5695674
#>  [6,] 0.7711601 0.7952634 0.5094968 0.20744615 0.7258932
#>  [7,] 0.7628150 1.2317141 0.8527672 0.46685097 1.0686435
#>  [8,] 0.0000000 1.7400265 1.3610796 0.97516334 1.5769559
#>  [9,] 1.7400265 0.0000000 0.2040745 0.44858546 0.5561065
#> [10,] 1.3610796 0.2040745 0.0000000 0.38005204 0.5499711
#> [11,] 0.9751633 0.4485855 0.3800520 0.00000000 0.4995983
#> [12,] 1.5769559 0.5561065 0.5499711 0.49959825 0.0000000
#> 
generate_random_model(d, 'general')
#> $graph
#> IGRAPH 515d760 U--- 12 11 -- Erdos-Renyi (gnp) graph
#> + attr: name (g/c), type (g/c), loops (g/l), p (g/n)
#> + edges from 515d760:
#>  [1]  2-- 5  3-- 5  5-- 7  6-- 7  4-- 8  6-- 8  4-- 9  1--11  6--11  8--12
#> [11] 10--12
#> 
#> $Gamma
#>            [,1]       [,2]        [,3]      [,4]        [,5]      [,6]
#>  [1,] 0.0000000  1.2308538  0.87786706 0.7584570  0.83480760 0.2328692
#>  [2,] 1.2308538  0.0000000  0.43910567 1.5235724  0.39604621 0.9979846
#>  [3,] 0.8778671  0.4391057  0.00000000 1.1705857  0.04305946 0.6449979
#>  [4,] 0.7584570  1.5235724  1.17058570 0.0000000  1.12752623 0.5255878
#>  [5,] 0.8348076  0.3960462  0.04305946 1.1275262  0.00000000 0.6019384
#>  [6,] 0.2328692  0.9979846  0.64499786 0.5255878  0.60193840 0.0000000
#>  [7,] 0.5072249  0.7236290  0.37064221 0.7999435  0.32758274 0.2743557
#>  [8,] 0.3647879  1.1299033  0.77691658 0.3936691  0.73385711 0.1319187
#>  [9,] 0.9042390  1.6693544  1.31636766 0.1457820  1.27330819 0.6713698
#> [10,] 9.9342146 10.6993300 10.34634327 9.9630958 10.30328380 9.7013454
#> [11,] 0.2167794  1.0140744  0.66108766 0.5416776  0.61802820 0.0160898
#> [12,] 0.5028813  1.2679967  0.91500992 0.5317625  0.87195046 0.2700121
#>            [,7]      [,8]       [,9]     [,10]     [,11]     [,12]
#>  [1,] 0.5072249 0.3647879  0.9042390  9.934215 0.2167794 0.5028813
#>  [2,] 0.7236290 1.1299033  1.6693544 10.699330 1.0140744 1.2679967
#>  [3,] 0.3706422 0.7769166  1.3163677 10.346343 0.6610877 0.9150099
#>  [4,] 0.7999435 0.3936691  0.1457820  9.963096 0.5416776 0.5317625
#>  [5,] 0.3275827 0.7338571  1.2733082 10.303284 0.6180282 0.8719505
#>  [6,] 0.2743557 0.1319187  0.6713698  9.701345 0.0160898 0.2700121
#>  [7,] 0.0000000 0.4062744  0.9457254  9.975701 0.2904455 0.5443677
#>  [8,] 0.4062744 0.0000000  0.5394511  9.569427 0.1480085 0.1380933
#>  [9,] 0.9457254 0.5394511  0.0000000 10.108878 0.6874596 0.6775444
#> [10,] 9.9757011 9.5694267 10.1088778  0.000000 9.7174352 9.4313333
#> [11,] 0.2904455 0.1480085  0.6874596  9.717435 0.0000000 0.2861019
#> [12,] 0.5443677 0.1380933  0.6775444  9.431333 0.2861019 0.0000000
#>