Conversion between Huesler-Reiss parameter matrices

Converts between different matrices that parametrize the same Huesler-Reiss distribution: , , , , . The 1 matrices and can also be given/returned as matrices with the kth row and column filled with zeros.

Usage

Gamma2Sigma(Gamma, k = NULL, full = FALSE, check = TRUE)

Gamma2Theta(Gamma, k = NULL, full = FALSE, check = TRUE)

Sigma2Gamma(Sigma, k = NULL, full = FALSE, check = TRUE)

Theta2Gamma(Theta, k = NULL, full = FALSE, check = TRUE)

Sigma2Theta(
  Sigma,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Theta2Sigma(
  Theta,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Theta2Theta(
  Theta,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Sigma2Sigma(
  Sigma,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Gamma2Gamma(Gamma, check = TRUE)

matrix2matrix(
  M,
  name1 = c("Gamma", "Sigma", "Theta")[1],
  name2 = c("Gamma", "Sigma", "Theta")[1],
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Arguments

Gamma

Numeric variogram matrix.

k

NULL if the input/output matrix is /. Else, an integer between 1 and d indicating the value of k in , .

full

Logical. If TRUE and !is.null(k), the input/output matrix is a matrix with the kth row filled with zeros.

check

Whether to check the inputs and call ensure_matrix_symmetry_and_truncate_zeros on the outputs.

Sigma

Numeric or 1 covariance matrix.

Theta

Numeric or 1 precision matrix.

k1

NULL if the input matrix is /. Else, an integer between 1 and d indicating the value of k in , .

k2

NULL if the output matrix is /. Else, an integer between 1 and d indicating the value of k in , .

full1

Logical. If TRUE and !is.null(k1), the input is a matrix with the kth row filled with zeros.

full2

Logical. If TRUE and !is.null(k2), the output is a matrix with the kth row filled with zeros.

M

Numeric matrix, , , or .

name1

Name of the input representation.

name2

Name of the output representation.

Value

The desired parameter matrix corresponding to the specified inputs.

Details

If k, k1, or k2 is NULL, the corresponding full* argument is ignored.

Gamma2Gamma only checks and returns the input.

matrix2matrix is a wrapper function that calls the corresponding conversion function implied by name1, name2.

See also