Sampling of a multivariate Pareto distribution

Simulates exact samples of a multivariate Pareto distribution.

rmpareto(n, model = c("HR", "logistic", "neglogistic", "dirichlet")[1], d, par)

Arguments

n
Number of simulations.

model
The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d
Dimension of the multivariate Pareto distribution.

par
Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric matrix of size \(n \times d\) of simulations of the multivariate Pareto distribution.

Details

The simulation follows the algorithm in Engelke and Hitz (2020) . For details on the parameters of the Huesler--Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles SG, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmpareto(n, "HR", d = d, par = G)

#>            [,1]      [,2]       [,3]      [,4]
#>  [1,] 1.5417348 0.6687125 0.53179013 0.9310867
#>  [2,] 0.3319329 1.1119546 0.18435378 0.9126737
#>  [3,] 6.3837104 3.0451475 2.35883930 5.6052171
#>  [4,] 1.3956915 1.5677871 2.90441830 2.0780491
#>  [5,] 0.9775273 1.3798756 0.04752097 2.4238717
#>  [6,] 0.4705918 0.6246992 0.50328369 2.3639409
#>  [7,] 1.1608287 0.9936025 0.20393843 1.1606914
#>  [8,] 1.0411925 0.1890028 0.92558369 0.1272640
#>  [9,] 1.0161098 0.1093052 0.08739437 0.3216439
#> [10,] 0.5842568 1.1467366 0.04405890 0.1015256

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmpareto(n, "logistic", d, par = theta)

#>             [,1]       [,2]       [,3]
#>  [1,] 1.07713867 1.62375233 1.65119930
#>  [2,] 0.20357793 1.06984474 0.50643524
#>  [3,] 1.21457338 0.05300082 0.24811185
#>  [4,] 1.42037615 0.79059451 0.46751597
#>  [5,] 0.34949212 2.18362845 0.97523543
#>  [6,] 0.39631663 2.13184020 1.98002940
#>  [7,] 0.20531414 1.17498618 0.18683788
#>  [8,] 0.17424730 2.86248339 0.04829661
#>  [9,] 0.04228709 0.05728603 1.06836094
#> [10,] 3.70759358 1.81749352 1.98290418

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmpareto(n, "neglogistic", d, par = theta)

#>            [,1]      [,2]        [,3]      [,4]      [,5]
#>  [1,] 1.8312599 1.1056142  0.16803787 2.8631817 3.9601745
#>  [2,] 0.5663816 1.0546481  0.56278179 0.7307199 0.8226854
#>  [3,] 1.1603439 0.4843039  0.47790658 0.6713704 0.4976342
#>  [4,] 0.5328217 0.8339900  1.30645360 0.3879871 0.5948820
#>  [5,] 5.9972664 1.5678082  1.30669219 2.7786683 2.5929902
#>  [6,] 0.2882552 0.5710630  0.17426849 1.4431127 0.8536229
#>  [7,] 6.2031360 2.8220883 10.10944457 2.3214239 4.7585770
#>  [8,] 0.7684642 1.6845776  0.43974936 0.3676126 1.2205655
#>  [9,] 0.8909148 1.0072556  0.09266287 0.5188209 1.0250493
#> [10,] 4.1313549 5.6845342  1.15597328 3.1637688 3.9774906

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmpareto(n, "dirichlet", d, par = alpha)

#>             [,1]       [,2]         [,3]       [,4]
#>  [1,] 0.25289887 9.99416098 0.0177259816 1.97780084
#>  [2,] 1.09896983 1.49320476 0.3114441555 0.65700256
#>  [3,] 0.26284400 1.40961075 0.0002279827 1.44103791
#>  [4,] 0.99112753 9.81882537 2.0691669183 4.24474822
#>  [5,] 0.53112063 0.01839776 2.0969161208 0.05385129
#>  [6,] 0.28346248 1.05827208 0.1787620633 0.48736746
#>  [7,] 8.12293367 5.56926884 4.0087540295 1.87097983
#>  [8,] 2.85776326 0.34059646 1.0899375411 0.98070566
#>  [9,] 3.72499394 0.28038407 0.0379689534 0.16903621
#> [10,] 0.08246743 0.15987726 1.4780748949 1.41080036