Sampling of a multivariate Pareto distribution

Simulates exact samples of a multivariate Pareto distribution.

Usage

rmpareto(
  n,
  model = c("HR", "logistic", "neglogistic", "dirichlet"),
  d = NULL,
  par
)

Arguments

n

Number of simulations.

model

The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d

Dimension of the multivariate Pareto distribution. In some cases this can be NULL and will be inferred from par.

par

Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric \(n \times d\) matrix of simulations of the multivariate Pareto distribution.

Details

The simulation follows the algorithm in Engelke and Hitz (2020) . For details on the parameters of the Huesler-Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles S, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

See also

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmpareto(n, "HR", d = d, par = G)
#>               [,1]         [,2]         [,3]        [,4]
#>  [1,] 4.615955e-02 6.271463e-02 4.679932e-01   1.5407978
#>  [2,] 1.195558e+00 6.472809e-01 5.718531e-01   0.4792028
#>  [3,] 1.819392e+00 3.652545e-01 3.662150e+00   0.3761302
#>  [4,] 2.045067e-01 1.301331e+00 5.665407e-02   0.1698795
#>  [5,] 4.908828e-01 1.173293e-01 2.697708e-01   1.2007516
#>  [6,] 1.896467e+00 3.308719e-01 9.912923e-02   0.1263963
#>  [7,] 3.783434e+03 7.740369e+03 1.928350e+03 955.7516474
#>  [8,] 4.075726e-02 1.452662e-01 2.243282e-01   1.3937023
#>  [9,] 1.001874e+00 4.463799e-01 6.635540e-02   0.1160326
#> [10,] 3.485334e+00 3.876919e+00 5.173686e+00   3.9681296

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmpareto(n, "logistic", d, par = theta)
#>             [,1]        [,2]       [,3]
#>  [1,] 2.49378501  0.36932494 0.33515179
#>  [2,] 0.99676085  0.83357935 2.32840596
#>  [3,] 1.46294810  0.39503110 0.16716777
#>  [4,] 0.04376721  0.07510183 1.41969112
#>  [5,] 1.36830518  0.67856567 0.65795762
#>  [6,] 1.19275582  9.60772370 0.76057218
#>  [7,] 2.09232247  6.71823647 1.09297062
#>  [8,] 1.03106621  0.14916308 0.19576513
#>  [9,] 0.09086568  1.77468972 0.07800615
#> [10,] 0.90695285 10.56331525 0.77231856

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmpareto(n, "neglogistic", d, par = theta)
#>            [,1]      [,2]      [,3]      [,4]      [,5]
#>  [1,] 7.3511435 3.9328194 2.6864731 6.1845938 0.4765336
#>  [2,] 1.8865512 3.3529559 7.2595824 1.0236510 3.1863134
#>  [3,] 4.0757165 0.7668703 1.4644710 7.0899404 1.9207669
#>  [4,] 7.9898612 8.1998763 2.6566668 5.5068894 1.6494460
#>  [5,] 1.0975102 1.3297490 0.9195467 0.2603207 2.3212560
#>  [6,] 0.5160742 0.7794853 0.3982861 1.2183559 1.9610998
#>  [7,] 0.1844014 1.0008575 0.1536657 0.6036499 0.2900357
#>  [8,] 1.0448853 0.2645238 0.1226914 0.1486905 0.7434475
#>  [9,] 2.1672959 0.2158939 1.9209722 0.3735742 1.2956526
#> [10,] 1.6389600 0.7371918 1.3493068 2.1050342 0.8566426

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmpareto(n, "dirichlet", d, par = alpha)
#>             [,1]      [,2]       [,3]      [,4]
#>  [1,] 0.55069296 0.1046215 3.08883131 0.3686149
#>  [2,] 0.86366602 0.2726818 0.15910693 1.1708847
#>  [3,] 0.39398663 0.1560021 3.77810676 2.2249727
#>  [4,] 1.67224085 1.7114601 0.06011546 2.4097557
#>  [5,] 0.07655213 1.2815127 0.05304225 0.3487650
#>  [6,] 9.99284057 0.7795481 0.26850632 3.3858871
#>  [7,] 2.17609618 0.4388302 0.27109554 2.3983842
#>  [8,] 0.38166196 0.1797033 1.31678931 0.1632469
#>  [9,] 0.30408976 0.3562097 1.11330902 0.6831740
#> [10,] 0.19853537 1.2816536 0.12636718 0.8424658