Sampling of a multivariate Pareto distribution

Simulates exact samples of a multivariate Pareto distribution.

Usage

rmpareto(n, model = c("HR", "logistic", "neglogistic", "dirichlet"), d, par)

Arguments

n

Number of simulations.

model

The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d

Dimension of the multivariate Pareto distribution.

par

Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric \(n \times d\) matrix of simulations of the multivariate Pareto distribution.

Details

The simulation follows the algorithm in Engelke and Hitz (2020) . For details on the parameters of the Huesler--Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles SG, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmpareto(n, "HR", d = d, par = G)
#>             [,1]       [,2]       [,3]       [,4]
#>  [1,] 16.3548939 51.8914647 26.5360057 11.9100801
#>  [2,] 20.9130951 32.0792224  7.2794536  5.6055372
#>  [3,]  1.2241391  0.3447070  0.7496115  0.2745461
#>  [4,]  0.7394802  0.4819105  2.1271742  0.3886519
#>  [5,]  1.0288000  0.6176337  4.4626858  3.2288565
#>  [6,]  4.3463725  2.2569587  1.5039667  0.4394725
#>  [7,]  0.3639499  0.5788987  0.1448434  1.0622063
#>  [8,]  0.2240812  0.7406062  1.6868892  1.0045579
#>  [9,]  2.5227895  2.7013269  1.2707881  6.1357386
#> [10,]  5.0506833  0.6972832  2.9392381  0.7087562

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmpareto(n, "logistic", d, par = theta)
#>              [,1]       [,2]       [,3]
#>  [1,]  0.05348099 1.43118239 0.05389678
#>  [2,]  0.01689916 0.01781006 1.99043774
#>  [3,]  6.44522829 3.84192615 4.75185297
#>  [4,] 11.15086927 0.16993262 0.24024980
#>  [5,]  2.25358425 4.15025057 0.98266205
#>  [6,]  0.79767190 1.01037069 0.22238741
#>  [7,]  2.78516354 0.21910546 1.20811417
#>  [8,]  0.04791265 0.10796482 1.68464391
#>  [9,]  0.30041421 0.34682216 3.42222481
#> [10,]  1.07739392 1.31927799 3.78082996

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmpareto(n, "neglogistic", d, par = theta)
#>            [,1]       [,2]      [,3]      [,4]       [,5]
#>  [1,] 0.9134682  0.4193091 0.7498344 1.7362368 0.08409103
#>  [2,] 0.7208462  3.3963208 1.7810699 0.3657981 2.16306778
#>  [3,] 1.3045793  2.5554570 0.5059166 1.9259453 1.72207737
#>  [4,] 0.2339351  0.4125991 0.8470116 1.2650381 0.42853321
#>  [5,] 0.2743498  1.2312591 0.5132162 0.4067593 0.07696845
#>  [6,] 1.1723923 15.6433440 2.5125949 4.3020861 1.53768302
#>  [7,] 0.2016467  0.8811418 1.5944097 0.9142069 0.75295434
#>  [8,] 1.1345907 13.6006050 7.5408809 1.6431343 0.70043126
#>  [9,] 1.3882851  1.3071895 0.1377822 0.7802426 0.35501419
#> [10,] 0.7799429  1.2795155 1.3238882 4.6093305 1.16572037

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmpareto(n, "dirichlet", d, par = alpha)
#>             [,1]      [,2]       [,3]      [,4]
#>  [1,] 0.57371501 1.2714518 0.05502085 0.4254865
#>  [2,] 0.27543374 1.1791533 0.71491748 0.1394769
#>  [3,] 0.18748175 1.0704765 0.20897389 0.2447078
#>  [4,] 0.03798897 0.5333116 0.77020255 1.0527140
#>  [5,] 5.90239676 7.4359643 3.69547405 9.2465348
#>  [6,] 1.64891085 0.7637040 0.73301231 2.4138431
#>  [7,] 0.22732093 4.1917728 0.31475693 1.2420307
#>  [8,] 0.55205514 0.3883455 1.07991041 0.2559198
#>  [9,] 0.40841575 0.3063875 0.73968876 1.0118474
#> [10,] 0.05650611 0.9281560 1.79495680 0.5093663