Sampling of a multivariate Pareto distribution

Simulates exact samples of a multivariate Pareto distribution.

rmpareto(n, model = c("HR", "logistic", "neglogistic", "dirichlet")[1], d, par)

Arguments

n
Number of simulations.

model
The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d
Dimension of the multivariate Pareto distribution.

par
Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric matrix of size \(n \times d\) of simulations of the multivariate Pareto distribution.

Details

The simulation follows the algorithm in Engelke and Hitz (2020) . For details on the parameters of the Huesler--Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles SG, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmpareto(n, "HR", d = d, par = G)

#>             [,1]       [,2]        [,3]       [,4]
#>  [1,]  3.4614684  5.2898207   2.4758344 12.1646286
#>  [2,]  2.2883341  0.3247503   1.2845164  2.7450237
#>  [3,]  1.0443831  0.7536959   2.1761806  0.1349619
#>  [4,]  1.2780636  7.2077891   0.6122561  4.2762275
#>  [5,] 12.8480579 24.1198879 133.7224687 77.1704660
#>  [6,] 10.4237737  2.0740572   3.0641864  3.2412732
#>  [7,]  1.2454707  4.3413699   1.0568667  1.3022388
#>  [8,] 16.8315718 91.5076379  46.4733485  5.1676352
#>  [9,]  0.8815399  1.4422442   0.8578199  0.9625599
#> [10,] 14.6369061  6.3426891   4.0847542  2.1925315

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmpareto(n, "logistic", d, par = theta)

#>             [,1]       [,2]      [,3]
#>  [1,]  1.1458283 1.31571280 0.6234862
#>  [2,]  0.6573020 1.34330352 0.8404121
#>  [3,]  2.3532707 0.09811965 0.2874037
#>  [4,] 12.0133901 2.89578077 2.3545052
#>  [5,]  0.2334423 0.87876020 1.0591711
#>  [6,]  0.5940259 1.32240821 0.7226196
#>  [7,]  0.9559918 1.18442247 1.8737507
#>  [8,]  0.5325738 0.29962265 1.3528127
#>  [9,]  2.3691356 0.38911612 0.4174353
#> [10,]  1.9418421 2.09868691 1.8375754

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmpareto(n, "neglogistic", d, par = theta)

#>            [,1]       [,2]      [,3]       [,4]      [,5]
#>  [1,] 5.1037100  2.9638609 1.6119464 3.36268309 0.4920120
#>  [2,] 0.3922700  0.3118590 7.5977256 0.50181040 1.8713482
#>  [3,] 1.6482697  0.6027669 0.7422611 1.09201001 0.9434034
#>  [4,] 1.6515694 10.8071958 4.8562936 4.55670855 1.0040608
#>  [5,] 0.2321307  0.1789395 0.9613249 1.05839578 0.8652831
#>  [6,] 0.7983551  0.4826561 2.3642042 3.32478267 0.9024587
#>  [7,] 0.1034105  1.1950463 0.2541052 0.41839079 0.4236146
#>  [8,] 0.3950065  4.2805655 2.9594218 1.77322301 1.9105707
#>  [9,] 0.1585603  1.0197425 0.7678065 0.18714017 0.4309584
#> [10,] 0.2039187  1.2768189 1.3892443 0.08161092 0.5831107

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmpareto(n, "dirichlet", d, par = alpha)

#>              [,1]         [,2]       [,3]        [,4]
#>  [1,]  2.88962213 1.836035e+00 2.94325575  2.55045246
#>  [2,] 56.80644557 6.338229e-01 0.71509332 11.91634069
#>  [3,]  0.02812321 5.259400e-02 1.00623115  0.26572451
#>  [4,]  0.03956225 1.032318e+00 0.07847292  0.03858923
#>  [5,]  6.04079057 8.207812e+00 1.71193864  3.61139563
#>  [6,]  4.92327949 2.798387e+00 0.74309148  4.36651445
#>  [7,]  0.46855947 4.220309e+00 0.60896335  0.87966979
#>  [8,]  0.01309099 2.406269e-01 1.31245680  0.03271308
#>  [9,]  5.37993463 1.806724e+00 1.77823179  2.17179564
#> [10,]  0.30191867 3.882559e-05 1.40936368  0.31382578