Sampling of a multivariate Pareto distribution

Simulates exact samples of a multivariate Pareto distribution.

rmpareto(n, model = c("HR", "logistic", "neglogistic", "dirichlet")[1], d, par)

Arguments

n
Number of simulations.

model
The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d
Dimension of the multivariate Pareto distribution.

par
Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric matrix of size \(n \times d\) of simulations of the multivariate Pareto distribution.

Details

The simulation follows the algorithm in Engelke and Hitz (2020) . For details on the parameters of the Huesler--Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles SG, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmpareto(n, "HR", d = d, par = G)

#>            [,1]       [,2]       [,3]       [,4]
#>  [1,] 0.1806210 1.24879491 1.84044132 1.22224274
#>  [2,] 2.9616589 3.40031061 0.54446724 3.05017769
#>  [3,] 0.5554720 1.17566836 0.19236004 0.26067756
#>  [4,] 1.6630351 0.20138352 0.74669996 0.09409008
#>  [5,] 5.7147166 0.57532419 0.85257924 0.16713476
#>  [6,] 0.6024294 1.39066406 0.09368122 0.10992752
#>  [7,] 0.2191467 1.18785861 0.30722450 0.58207776
#>  [8,] 0.4267704 0.17243429 0.37255772 2.05458269
#>  [9,] 3.7376219 2.40759245 0.56933549 0.13608085
#> [10,] 0.1300231 0.04217159 1.44243083 0.91085218

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmpareto(n, "logistic", d, par = theta)

#>             [,1]       [,2]       [,3]
#>  [1,] 0.91562228 4.23158713 0.22787871
#>  [2,] 0.08773685 0.15241525 1.56190248
#>  [3,] 1.06835884 0.28190144 0.09542060
#>  [4,] 2.16361469 0.31879194 2.71293538
#>  [5,] 0.58543442 0.67168346 2.78296264
#>  [6,] 0.21341491 0.23912888 1.86300204
#>  [7,] 3.20255130 0.05418079 0.02204855
#>  [8,] 2.61315055 0.85576639 0.46957745
#>  [9,] 1.71818974 0.03715810 0.01550446
#> [10,] 2.95436002 5.76648563 2.98263223

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmpareto(n, "neglogistic", d, par = theta)

#>            [,1]      [,2]      [,3]       [,4]      [,5]
#>  [1,] 2.8397099 3.3896852 1.3432972 3.28837364 6.3260122
#>  [2,] 0.1095998 0.9721741 1.2544572 1.68206688 1.5340214
#>  [3,] 3.4693555 0.9642763 0.9870767 3.73204409 3.4185847
#>  [4,] 0.3578167 0.5021830 1.1808499 1.07655685 1.2180390
#>  [5,] 2.6379298 1.5884161 1.2833066 1.27033411 0.2166212
#>  [6,] 1.1171031 0.8619950 0.8662016 0.04194123 0.4500053
#>  [7,] 1.0405670 1.8556893 1.0739551 1.24317866 1.1859019
#>  [8,] 1.8578107 0.9972544 4.8087224 0.94668052 2.1130311
#>  [9,] 1.0034209 1.5091466 0.9789727 0.17226830 0.7708232
#> [10,] 0.3417708 2.9345180 0.6904892 2.13341286 1.3786231

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmpareto(n, "dirichlet", d, par = alpha)

#>              [,1]      [,2]       [,3]      [,4]
#>  [1,]  0.96800164 3.4625758 1.34886406 6.0516500
#>  [2,]  1.20982694 0.1276342 1.94137287 0.2313889
#>  [3,]  2.33831757 1.6231604 0.03516146 1.3044911
#>  [4,]  0.39049459 0.5568097 5.73611532 0.6102139
#>  [5,] 25.44012313 6.4199506 0.29009134 5.8416501
#>  [6,]  2.05777076 1.1414781 8.21296420 0.9475164
#>  [7,]  8.59530505 4.4426969 2.86495545 1.8890830
#>  [8,]  0.04858485 1.0665172 0.24619486 1.0020749
#>  [9,]  1.37011633 0.4693763 0.82894380 1.8550304
#> [10,]  1.03700500 0.1254493 0.52958008 0.8860396