Sampling of a multivariate Pareto distribution

Simulates exact samples of a multivariate Pareto distribution.

Usage

rmpareto(
  n,
  model = c("HR", "logistic", "neglogistic", "dirichlet"),
  d = NULL,
  par
)

Arguments

n

Number of simulations.

model

The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d

Dimension of the multivariate Pareto distribution. In some cases this can be NULL and will be inferred from par.

par

Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric \(n \times d\) matrix of simulations of the multivariate Pareto distribution.

Details

The simulation follows the algorithm in Engelke and Hitz (2020) . For details on the parameters of the Huesler-Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles S, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

Engelke S, Hitz AS (2020). “Graphical models for extremes (with discussion).” J. R. Stat. Soc. Ser. B Stat. Methodol., 82, 871--932.

See also

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmpareto(n, "HR", d = d, par = G)
#>             [,1]       [,2]        [,3]       [,4]
#>  [1,]  0.6242115  1.2340786  0.28424553  0.4945530
#>  [2,]  7.7306138  0.5739548  5.36225882  0.9660429
#>  [3,]  6.3546256 12.9548933  4.73217785  4.2658041
#>  [4,]  0.6558634  1.5353209  0.49389134  0.5448121
#>  [5,] 42.8554548  8.5273448 31.33763315 55.9390571
#>  [6,]  1.6246369  0.1199755  2.05703805  0.6370474
#>  [7,]  1.1139422  0.2996312  0.36922986  0.8431961
#>  [8,]  1.3597143  1.6552748  5.86433799  1.1239351
#>  [9,]  1.6800225  0.2931093  0.08781558  0.1119706
#> [10,]  0.6128817  1.2130212  0.90993956  0.4373595

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmpareto(n, "logistic", d, par = theta)
#>             [,1]        [,2]      [,3]
#>  [1,] 8.24837685 10.69857583 2.4734847
#>  [2,] 1.14082522  0.58671413 1.2991381
#>  [3,] 1.05199869  0.01952548 0.1518117
#>  [4,] 0.09815916  2.09836473 0.9171968
#>  [5,] 1.44549755  0.03732527 0.1230960
#>  [6,] 0.87907805  2.96681227 1.6638072
#>  [7,] 0.25061503  1.47779913 0.2151474
#>  [8,] 1.05973104  0.40336846 0.1357882
#>  [9,] 0.21627479  1.27101068 0.5378615
#> [10,] 0.76093562  0.58420217 1.3596907

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmpareto(n, "neglogistic", d, par = theta)
#>             [,1]      [,2]      [,3]      [,4]       [,5]
#>  [1,]  1.0410953 1.0368430 3.0641360 0.4899866  0.3413277
#>  [2,]  0.8043985 1.4684191 0.3371993 2.4377260  2.0067296
#>  [3,]  0.8398382 1.3174039 0.3297471 1.1210441  0.7363261
#>  [4,]  1.6320666 0.4856634 0.2237380 0.1971754  0.6634043
#>  [5,]  1.6686381 0.2726166 0.7718576 0.7949440  0.2269781
#>  [6,]  0.7743841 1.7759363 4.0861415 0.3360490  3.0394581
#>  [7,]  0.8424968 0.8434586 1.6880089 0.2612004  0.4319269
#>  [8,]  4.2622232 5.1144089 1.3030681 5.3718415  1.0723264
#>  [9,] 10.4418632 3.3544738 5.1462313 2.2192071 17.9442460
#> [10,]  1.1775948 0.5574875 1.1206045 6.1359100  1.2798100

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmpareto(n, "dirichlet", d, par = alpha)
#>              [,1]       [,2]       [,3]      [,4]
#>  [1,]  0.54438164 0.03495636 0.89308764 1.6750301
#>  [2,]  4.66487201 0.28122718 2.04269019 1.5075597
#>  [3,]  2.60696064 1.67971723 1.70213708 1.5252516
#>  [4,]  0.07793149 0.38886068 1.15564558 0.1560207
#>  [5,]  2.97816809 0.28430839 0.77949134 1.9339135
#>  [6,]  1.44739369 1.33222612 0.01962053 0.3117005
#>  [7,]  1.08362319 0.32159735 0.52307518 0.1984790
#>  [8,]  0.03959297 0.01409648 0.60053325 4.2018204
#>  [9,] 17.73588870 0.15360482 9.95951584 8.4454189
#> [10,]  4.03849379 0.94996230 1.76963146 1.4442792