Sampling of a multivariate max-stable distribution

Simulates exact samples of a multivariate max-stable distribution.

Usage

rmstable(n, model = c("HR", "logistic", "neglogistic", "dirichlet")[1], d, par)

Arguments

n

Number of simulations.

model

The parametric model type; one of:

HR (default),

logistic,

neglogistic,

dirichlet.

d

Dimension of the multivariate Pareto distribution.

par

Respective parameter for the given model, that is,

\(\Gamma\), numeric \(d \times d\) variogram matrix, if model = HR.

\(\theta \in (0, 1)\), if model = logistic.

\(\theta > 0\), if model = neglogistic.

\(\alpha\), numeric vector of size d with positive entries, if model = dirichlet.

Value

Numeric \(n \times d\) matrix of simulations of the multivariate max-stable distribution.

Details

The simulation follows the extremal function algorithm in Dombry et al. (2016) . For details on the parameters of the Huesler-Reiss, logistic and negative logistic distributions see Dombry et al. (2016) , and for the Dirichlet distribution see Coles and Tawn (1991) .

References

Coles S, Tawn JA (1991). “Modelling extreme multivariate events.” J. R. Stat. Soc. Ser. B Stat. Methodol., 53, 377--392.

Dombry C, Engelke S, Oesting M (2016). “Exact simulation of max-stable processes.” Biometrika, 103, 303--317.

See also

Examples

## A 4-dimensional HR distribution
n <- 10
d <- 4
G <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

rmstable(n, "HR", d = d, par = G)
#>             [,1]       [,2]       [,3]       [,4]
#>  [1,]  0.9129700  2.7320869  4.3930684  2.0396394
#>  [2,]  0.2666331  0.7563013  1.6998662  0.7642649
#>  [3,]  1.4469430  3.2588614  0.6616103  0.9099434
#>  [4,]  0.4459274  1.6260210  0.6936030  1.0758879
#>  [5,]  0.4388372  0.7088128  0.4051627  0.5354991
#>  [6,]  0.5635592  0.6925003  2.5880064  0.4623138
#>  [7,]  1.6316665  1.2772479  2.3119953  6.8583689
#>  [8,] 25.5877634  9.4518333 30.2610583 11.8954021
#>  [9,]  9.1793281  9.6263068  7.7595231 18.7535580
#> [10,] 23.4589286 30.4513683 27.9736059 12.6065325

## A 3-dimensional logistic distribution
n <- 10
d <- 3
theta <- .6
rmstable(n, "logistic", d, par = theta)
#>            [,1]       [,2]      [,3]
#>  [1,] 0.7920967  0.3838593 0.5446596
#>  [2,] 0.4762412  1.7332709 1.3834479
#>  [3,] 1.3370597  0.2851836 4.0607764
#>  [4,] 0.7421880  1.8327530 2.2199731
#>  [5,] 1.9280030 10.7751455 1.5144490
#>  [6,] 1.2808886  0.8044145 0.8670885
#>  [7,] 1.2840544  0.7799514 5.0934757
#>  [8,] 0.5676049  0.3633312 0.3426581
#>  [9,] 0.2693890  0.3870772 0.4698337
#> [10,] 1.0291076  0.4829797 1.6501069

## A 5-dimensional negative logistic distribution
n <- 10
d <- 5
theta <- 1.5
rmstable(n, "neglogistic", d, par = theta)
#>             [,1]       [,2]       [,3]       [,4]       [,5]
#>  [1,]  7.6648654 12.7130188  5.1404650 29.9312790  6.7218769
#>  [2,]  0.6997510  0.7048263  1.2611360  0.8094214  0.5227736
#>  [3,] 23.7136939  1.8638248  7.7195594  8.6375029 18.3126055
#>  [4,]  1.1596061  0.9826885  1.1413147  1.3009600  0.4160243
#>  [5,]  1.3697433  0.6338478  0.7218860  1.4924707  0.9762449
#>  [6,]  2.2482129  2.8355838  1.4315486  4.9443567  1.6710074
#>  [7,] 16.7420023 15.0632920 12.9267446 20.3789551 28.2010827
#>  [8,]  0.4362455  0.3568023  0.3109209  0.9655873  0.3155885
#>  [9,]  0.8675428  0.6111543  0.4107973  0.8449127  0.6456791
#> [10,]  3.8236338  3.0036666  4.5845912  1.6076321  2.4079347

## A 4-dimensional Dirichlet distribution
n <- 10
d <- 4
alpha <- c(.8, 1, .5, 2)
rmstable(n, "dirichlet", d, par = alpha)
#>            [,1]      [,2]      [,3]      [,4]
#>  [1,] 1.8895508 0.7655561 1.2157059 2.1661873
#>  [2,] 0.7099777 1.8470676 3.6166997 0.7038762
#>  [3,] 3.0723931 0.4386526 0.9849517 0.7825669
#>  [4,] 0.5039515 0.4160592 0.9417922 2.8132326
#>  [5,] 8.8648594 2.9460901 7.5495707 6.8022550
#>  [6,] 2.8973698 4.9357845 2.1741180 4.9137229
#>  [7,] 1.5847216 6.1903019 3.3346973 1.8198441
#>  [8,] 0.2975653 1.2948784 1.5564330 2.1926874
#>  [9,] 1.3155831 0.5673063 0.4437899 1.0480459
#> [10,] 1.7603035 5.4723088 1.5448006 4.8258270